Tabulation problem when select value in select
When I select a new value and press tab button, the focus is not on the next field. This is due because the function initSelect is call all the time in wacher. To resolve it, check if the change of model is due to the select, and if the changement is due to select, they are nothing to do.